+++ game.js
@@ -74,6 +74,8 @@ var Game = (function() {
         self.quests.update(delta);
         self.questUI.update(delta);
         self.artifactUI.update(delta);
+        self.research.update(delta);
+        self.researchUI.update(delta);
 
         self.updateAutoSave(delta);
 
@@ -155,6 +157,7 @@ var Game = (function() {
         this.goldenEvents.save(data);
         this.quests.save(data);
         this.artifacts.save(data);
+        this.research.save(data);
         this.interstellar.save(data);
         this.stargaze.save(data);
         this.updates.save(data);
@@ -182,6 +185,7 @@ var Game = (function() {
             this.goldenEvents.load(data);
             this.quests.load(data);
             this.artifacts.load(data);
+            this.research.load(data);
             this.stargaze.load(data);
             this.tech.load(data);
             this.interstellar.load(data); 
@@ -268,6 +272,7 @@ var Game = (function() {
         self.goldenEvents.initialise();
         self.quests.initialise();
         self.artifacts.initialise();
+        self.research.initialise();
 
         // Now load
         self.load();
@@ -280,6 +285,7 @@ var Game = (function() {
         self.goldenEventUI.initialise();
         self.questUI.initialise();
         self.artifactUI.initialise();
+        self.researchUI.initialise();
 
         for(var i = 0; i < self.uiComponents.length; i++) {
             self.uiComponents[i].initialise();